requirements.txt
gradio
pandas
keras
tensorflow
sklearn
numpy